Move coordinate transformation to jinja
