+++ src/exsce_floorplan/templates/relations.json
@@ -1,0 +1,18 @@
+{# TODO Not sure why the graphs.helpers were needed #}
+{% macro from_ref_name(from_frame) -%}
+{%if from_frame.world %}
+world-frame
+{%- elif from_frame.ref is not none and from_frame.index %}
+frame-{{from_frame.ref.name}}-wall-{{from_frame.index}}
+{%- elif from_frame.ref is not none %}
+frame-center-{{from_frame.ref.name }}
+{%- endif %}
+{%- endmacro %}
+
+{% macro to_ref_name(to_frame, space) -%}
+{% if to_frame.index is not none %}
+frame-{{space.name}}-wall-{{to_frame.index}}
+{%- else %}
+frame-center-{{space.name}}
+{%- endif %}
+{%- endmacro %}
